3. Интерфейс Input [#181092 #179137]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -6,5 +6,5 @@ public interface Transport {
 
     public void passengers(int numberOfPassengers);
 
-    public int tankUp(int fuel);
+    public double tankUp(int fuel);
 }
